【hblog-view】将目录抽成一个组件
