+++ hblog-view/src/views/frontend/blog/BlogDetail.vue
@@ -12,20 +12,7 @@
             </div>
             <div class="affixContainer">
               <el-affix target=".affixContainer" :offset-top="0">
-                <el-card :class="{ 'hide-toc': hasToc }" style="margin-bottom: 14px;">
-                  <div class="card-title">目录</div>
-                  <div
-                      v-for="anchor in titles"
-                      :style="{ 'padding-left': `${anchor.indent * 20}px` }"
-                      :class="[
-                    'catalog-item',
-                    anchor.lineIndex === currentLineIndex ? 'active' : 'not-active',
-                ]"
-                      @click="handleAnchorClick(anchor)"
-                  >
-                    <a style="cursor: pointer">{{ anchor.title }}</a>
-                  </div>
-                </el-card>
+                <Catalog v-if="isLoaded" :titlesData="previewRef"/>
                 <RecommendBlogCard/>
               </el-affix>
             </div>
@@ -46,6 +33,7 @@
 <script lang="ts" setup>
 import PageCover from "@/views/frontend/layout/PageCover.vue"
 import OverviewCard from "@/components/OverviewCard.vue"
+import Catalog from "@/components/frontend/Catalog.vue"
 import RecommendBlogCard from "@/components/RecommendBlogCard.vue"
 import Footer from "@/views/frontend/layout/Footer.vue"
 // 博客数据
@@ -66,44 +54,11 @@ const getBlogDetailData = async () => {
   blog.value = res.body
 }
 
-let currentLineIndex = ref(null);
-
-const hasToc = ref(false);
-const loadToc = () => {
-  const preview = previewRef.value;
-  const anchors = preview.$el.querySelectorAll('h1,h2,h3,h4,h5,h6');
-  const titlesArray = Array.from(anchors).filter((title) => !!title.innerText.trim());
-  if (!titlesArray.length) {
-    titles.value = [];
-    hasToc.value = true;
-    return;
-  }
-  const hTags = Array.from(new Set(titlesArray.map((title) => title.tagName))).sort();
-  titles.value = titlesArray.map((el) => ({
-    title: el.innerText,
-    lineIndex: el.getAttribute('data-v-md-line'),
-    indent: hTags.indexOf(el.tagName),
-  }));
-}
-
-const handleAnchorClick = (anchor) => {
-  const preview = previewRef.value;
-  const {lineIndex} = anchor;
-  const heading = preview.$el.querySelector(`[data-v-md-line="${lineIndex}"]`);
-  currentLineIndex.value = anchor.lineIndex
-
-  if (heading) {
-    preview.scrollToTarget({
-      target: heading,
-      scrollContainer: window,
-      top: 60,
-    });
-  }
-};
+const isLoaded = ref(false);
 
 onMounted(() => {
   getBlogDetailData().then(() => {
-        loadToc()
+        isLoaded.value = true
       }
   )
 })
@@ -141,34 +96,6 @@ onMounted(() => {
   top: 20px;
 }
 
-.catalog-item {
-  color: #666261;
-  margin: 5px 0;
-  line-height: 28px;
-  cursor: pointer;
-  transition: all 0.2s ease-in-out;
-  font-size: 14px;
-  padding: 2px 6px;
-  display: -webkit-box;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  -webkit-line-clamp: 1;
-  -webkit-box-orient: vertical;
-
-  &:hover {
-    color: var(--theme-color);
-  }
-}
-
-.active {
-  background-color: var(--theme-color);
-  color: white;
-
-  &:hover {
-    background-color: #0c82e9;
-    color: white;
-  }
-}
 
 .sidebar {
   display: flex;
@@ -179,10 +106,6 @@ onMounted(() => {
 
 .sidebar-item {
   margin-bottom: 14px;
-}
-
-.hide-toc {
-  display: none;
 }
 
 @media screen and (max-width: 900px) {
